add test for print-defaults, move typer app to segmantic.commands
